Rename async interface to standard interface Rename sync interface to Sync...
